+++ pygdbmi/gdbmiparser.py
@@ -8,13 +8,33 @@ See more at https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI.html#GDB_002fMI
 
 """
 
-import re
-from pygdbmi.printcolor import print_green
+import logging
+from pygdbmi.printcolor import fmt_green
 from pygdbmi.StringStream import StringStream
 from pprint import pprint
+import re
 
-# Print text to console as it's being parsed to help debug
 _DEBUG = False
+logger = logging.getLogger(__name__)
+
+
+def _setup_logger(logger, debug):
+    logger.propagate = False
+
+    handler = logging.StreamHandler()
+    handler.setFormatter(
+        logging.Formatter("[%(filename)s:%(lineno)s - %(funcName)20s() ] %(message)s")
+    )
+    if debug:
+        level = logging.DEBUG
+    else:
+        level = logging.ERROR
+
+    logger.setLevel(level)
+    logger.addHandler(handler)
+
+
+_setup_logger(logger, _DEBUG)
 
 
 def parse_response(gdb_mi_text):
@@ -87,6 +107,7 @@ def response_is_finished(gdb_mi_text):
     Returns: True if gdb response is finished"""
     if _GDB_MI_RESPONSE_FINISHED_RE.match(gdb_mi_text):
         return True
+
     else:
         return False
 
@@ -116,35 +137,35 @@ def assert_match(actual_char_or_str, expected_char_or_str):
 # In addition to a number of out-of-band notifications,
 # the response to a gdb/mi command includes one of the following result indications:
 # done, running, connected, error, exit
-_GDB_MI_RESULT_RE = re.compile("^(\d*)\^(\S+?)(,(.*))?$")
+_GDB_MI_RESULT_RE = re.compile(r"^(\d*)\^(\S+?)(,(.*))?$")
 
 # https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Async-Records.html#GDB_002fMI-Async-Records
 # Async records are used to notify the gdb/mi client of additional
 # changes that have occurred. Those changes can either be a consequence
 # of gdb/mi commands (e.g., a breakpoint modified) or a result of target activity
 # (e.g., target stopped).
-_GDB_MI_NOTIFY_RE = re.compile("^(\d*)[*=](\S+?),(.*)$")
+_GDB_MI_NOTIFY_RE = re.compile(r"^(\d*)[*=](\S+?),(.*)$")
 
 # https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Stream-Records.html#GDB_002fMI-Stream-Records
 # "~" string-output
 # The console output stream contains text that should be displayed
 # in the CLI console window. It contains the textual responses to CLI commands.
-_GDB_MI_CONSOLE_RE = re.compile('~"(.*)"', re.DOTALL)
+_GDB_MI_CONSOLE_RE = re.compile(r'~"(.*)"', re.DOTALL)
 
 # https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Stream-Records.html#GDB_002fMI-Stream-Records
 # "&" string-output
 # The log stream contains debugging messages being produced by gdb's internals.
-_GDB_MI_LOG_RE = re.compile('&"(.*)"', re.DOTALL)
+_GDB_MI_LOG_RE = re.compile(r'&"(.*)"', re.DOTALL)
 
 # https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Stream-Records.html#GDB_002fMI-Stream-Records
 # "@" string-output
 # The target output stream contains any textual output from the
 # running target. This is only present when GDB's event loop is truly asynchronous,
 # which is currently only the case for remote targets.
-_GDB_MI_TARGET_OUTPUT_RE = re.compile('@"(.*)"', re.DOTALL)
+_GDB_MI_TARGET_OUTPUT_RE = re.compile(r'@"(.*)"', re.DOTALL)
 
 # Response finished
-_GDB_MI_RESPONSE_FINISHED_RE = re.compile("^\(gdb\)\s*$")
+_GDB_MI_RESPONSE_FINISHED_RE = re.compile(r"^\(gdb\)\s*$")
 
 _WHITESPACE = [" ", "\t", "\r", "\n"]
 
@@ -162,12 +183,11 @@ def _get_notify_msg_and_payload(result, stream):
     """Get notify message and payload dict"""
     token = stream.advance_past_chars(["=", "*"])
     token = int(token) if token != "" else None
-    if _DEBUG:
-        print_green("parsing message")
+    logger.debug("%s", fmt_green("parsing message"))
     message = stream.advance_past_chars([","])
-    if _DEBUG:
-        print("parsed message")
-        print_green(message)
+
+    logger.debug("parsed message")
+    logger.debug("%s", fmt_green(message))
 
     payload = _parse_dict(stream)
     return token, message.strip(), payload
@@ -197,8 +217,8 @@ def _parse_dict(stream):
     """
     obj = {}
 
-    if _DEBUG:
-        print_green("parsing dict")
+    logger.debug("%s", fmt_green("parsing dict"))
+
     while True:
         c = stream.read(1)
         if c in _WHITESPACE:
@@ -208,6 +228,7 @@ def _parse_dict(stream):
         elif c in ["}", ""]:
             # end of object, exit loop
             break
+
         else:
             stream.seek(-1)
             key, val = _parse_key_val(stream)
@@ -237,14 +258,12 @@ def _parse_dict(stream):
                     # got some garbage text, skip it. for example:
                     # name="gdb"gargage  # skip over 'garbage'
                     # name="gdb"\n  # skip over '\n'
-                    if _DEBUG:
-                        print("skipping unexpected charcter" + c)
+                    logger.debug("skipping unexpected charcter: " + c)
                     c = stream.read(1)
             stream.seek(-1)
 
-    if _DEBUG:
-        print_green("parsed dict")
-        print_green(obj)
+    logger.debug("parsed dict")
+    logger.debug("%s", fmt_green(obj))
     return obj
 
 
@@ -255,15 +274,14 @@ def _parse_key_val(stream):
         Parsed value (either a string, array, or dict)
     """
 
-    if _DEBUG:
-        print_green("parsing key/val")
+    logger.debug("parsing key/val")
     key = _parse_key(stream)
     val = _parse_val(stream)
 
-    if _DEBUG:
-        print_green("parsed key/val")
-        print_green(key)
-        print_green(val)
+    logger.debug("parsed key/val")
+    logger.debug("%s", fmt_green(key))
+    logger.debug("%s", fmt_green(val))
+
     return key, val
 
 
@@ -272,12 +290,12 @@ def _parse_key(stream):
     returns :
         Parsed key (string)
     """
-    if _DEBUG:
-        print_green("parsing key")
+    logger.debug("parsing key")
+
     key = stream.advance_past_chars(["="])
-    if _DEBUG:
-        print_green("parsed key:")
-        print_green(key)
+
+    logger.debug("parsed key:")
+    logger.debug("%s", fmt_green(key))
     return key
 
 
@@ -287,8 +305,7 @@ def _parse_val(stream):
         Parsed value (either a string, array, or dict)
     """
 
-    if _DEBUG:
-        print_green("parsing value")
+    logger.debug("parsing value")
 
     while True:
         c = stream.read(1)
@@ -297,16 +314,20 @@ def _parse_val(stream):
             # Start object
             val = _parse_dict(stream)
             break
+
         elif c == "[":
             # Start of an array
             val = _parse_array(stream)
             break
+
         elif c == '"':
             # Start of a string
             val = stream.advance_past_string_with_gdb_escapes()
             break
+
         elif _DEBUG:
             raise ValueError("unexpected character: %s" % c)
+
         else:
             print(
                 'pygdbmi warning: encountered unexpected character: "%s". Continuing.'
@@ -314,9 +335,8 @@ def _parse_val(stream):
             )
             val = ""  # this will be overwritten if there are more characters to be read
 
-    if _DEBUG:
-        print_green("parsed value:")
-        print_green(val)
+    logger.debug("parsed value:")
+    logger.debug("%s", fmt_green(val))
 
     return val
 
@@ -327,8 +347,7 @@ def _parse_array(stream):
         Parsed array
     """
 
-    if _DEBUG:
-        print_green("parsing array")
+    logger.debug("parsing array")
     arr = []
     while True:
         c = stream.read(1)
@@ -346,7 +365,6 @@ def _parse_array(stream):
             # that elements of this array can be also be arrays.
             break
 
-    if _DEBUG:
-        print("parsed array:")
-        print_green(arr)
+    logger.debug("parsed array:")
+    logger.debug("%s", fmt_green(arr))
     return arr