+++ tests/test_pygdbmi.py
@@ -26,7 +26,6 @@ else:
 
 
 class TestPyGdbMi(unittest.TestCase):
-
     def test_parser(self):
         """Test that the parser returns dictionaries from gdb mi strings as expected"""
 
@@ -157,6 +156,7 @@ class TestPyGdbMi(unittest.TestCase):
             os.path.dirname(os.path.realpath(__file__)), "sample_c_app"
         )
         binary_path = os.path.join(SAMPLE_C_CODE_DIR, binary_name)
+        subprocess.call(["rm", "pygdbmi.a*"], cwd=SAMPLE_C_CODE_DIR)
         # Build C program
         subprocess.check_output(
             [MAKE_CMD, makefile_target_name, "-C", SAMPLE_C_CODE_DIR, "--quiet"]
@@ -182,9 +182,8 @@ class TestPyGdbMi(unittest.TestCase):
         # Verify output was parsed into a list of responses
         assert len(responses) != 0
         response = responses[0]
-        assert set(response.keys()) == set(
-            ["message", "type", "payload", "stream", "token"]
-        )
+        assert set(response.keys()) == {"message", "type", "payload", "stream", "token"}
+
         assert response["message"] == "thread-group-added"
         assert response["type"] == "notify"
         assert response["payload"] == {"id": "i1"}
@@ -257,14 +256,13 @@ class TestPyGdbMi(unittest.TestCase):
         to_be_buffered = b'^done,BreakpointTable={nr_rows="1",nr_'
 
         stream = "teststream"
-        verbose = False
-        response = gdbmi._get_responses_list(to_be_buffered, stream, verbose)
+        response = gdbmi._get_responses_list(to_be_buffered, stream)
         # Nothing should have been parsed yet
         assert len(response) == 0
         assert gdbmi._incomplete_output[stream] == to_be_buffered
 
         remaining_gdb_output = b'cols="6"}\n(gdb) \n'
-        response = gdbmi._get_responses_list(remaining_gdb_output, stream, verbose)
+        response = gdbmi._get_responses_list(remaining_gdb_output, stream)
 
         # Should have parsed response at this point
         assert len(response) == 1
@@ -291,9 +289,10 @@ class TestPyGdbMi(unittest.TestCase):
                     gdb_mi_simulated_output = f.read(n)
                     if gdb_mi_simulated_output == b"":
                         break  # EOF
+
                     # let the controller try to parse this additional raw gdb output
                     responses += gdbmi._get_responses_list(
-                        gdb_mi_simulated_output, stream, False
+                        gdb_mi_simulated_output, stream
                     )
             assert len(responses) == 141
 
@@ -374,7 +373,6 @@ class TestPyGdbMi(unittest.TestCase):
 
 
 class TestPerformance(unittest.TestCase):
-
     def get_test_input(self, n_repetitions):
         data = ", ".join(
             ['"/a/path/to/parse/' + str(i) + '"' for i in range(n_repetitions)]
@@ -406,7 +404,6 @@ class TestPerformance(unittest.TestCase):
 
 
 class TestStringStream(unittest.TestCase):
-
     def test_api(self):
         raw_text = 'abc- "d" ""ef"" g'
         stream = StringStream(raw_text)
@@ -451,4 +448,4 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    exit(main())