+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # pygdbmi release history
 
+## 0.9.0.0
+* Stop buffering output
+* Use logger in GdbController; modify `verbose` arguments.
+* Remove support for Python 3.3
+
+## 0.8.4.0
+* Add method `get_subprocess_cmd` to view the gdb command run in the shell
+
 ## 0.8.3.0
 * Improve reading gdb responses on unix (performance, bugfix) (@mouuff)